add eslint tsconf

+++ db/db_users.ts
@@ -8,7 +8,7 @@ class DBBase {
 	public tokenCollection: Collection<Tokens>;
 	public client: MongoClient;
 	constructor() {
-		this.client = new MongoClient(process.env.DB_URI);
+		this.client = new MongoClient(String(process.env.DB_URI));
 		this.db = this.client.db(process.env.DB_NAME);
 		this.usersCollection = this.db.collection<User>("users");
 		this.tokenCollection = this.db.collection<Tokens>("tokens");
@@ -18,8 +18,8 @@ type UserWithTokens = User & { tokenDocument: Tokens[] };
 export class DBUsers extends DBBase {
 	async addUsers(users: User[]) {
 		try {
-			const usersIntoDb = [];
-			for (let user of users) {
+			const usersIntoDb: User[] = [];
+			for (const user of users) {
 				usersIntoDb.push(
 					new User(user.login, await user.getHashPassword(), user.is_admin)
 				);
@@ -27,6 +27,7 @@ export class DBUsers extends DBBase {
 			const createdUsers = await this.usersCollection.insertMany(usersIntoDb);
 			return Object.values(createdUsers.insertedIds);
 		} catch (e) {
+			console.log(e);
 		} finally {
 			await this.client.close();
 		}
@@ -36,19 +37,21 @@ export class DBUsers extends DBBase {
 			const usersDeleteFilter: Filter<User> =
 				users[0] instanceof User
 					? {
-							login: { $in: users.map((e) => e.login) },
-					  }
+						login: { $in: (users as User[]).map((e) => e.login) },
+					}
 					: { _id: { $in: users as ObjectId[] } };
-			let tokensDeleteFilter: Filter<Tokens> = null;
+			let tokensDeleteFilter: Filter<Tokens> | null = null;
 			if (users[0] instanceof User) {
 				const idsToDelete = await this.usersCollection
 					.find(usersDeleteFilter)
 					.toArray();
-				tokensDeleteFilter = { user: { $in: idsToDelete.map((e) => e._id) } };
+				tokensDeleteFilter = { user: { $in: (idsToDelete.map((e) => (e._id as ObjectId))) } };
 			} else {
 				tokensDeleteFilter = { user: { $in: users as ObjectId[] } };
 			}
-			await this.tokenCollection.deleteMany(tokensDeleteFilter);
+			if (tokensDeleteFilter) {
+				await this.tokenCollection.deleteMany(tokensDeleteFilter);
+			}
 			await this.usersCollection.deleteMany(usersDeleteFilter);
 		} catch (e) {
 			console.log(e);
@@ -58,28 +61,29 @@ export class DBUsers extends DBBase {
 	}
 
 	async updatedUserToken(data: {
-		user: ObjectId | User;
+		user: ObjectId | User | null;
 		accessToken: string;
 		refreshToken: string;
 	}) {
 		try {
 			const userId = data.user instanceof User ? data.user._id : data.user;
-			await this.tokenCollection.updateOne(
-				{ user: userId },
-				{
-					$set: {
-						access_token: data.accessToken,
-						refresh_token: data.refreshToken,
+			if (userId)
+				await this.tokenCollection.updateOne(
+					{ user: userId },
+					{
+						$set: {
+							access_token: data.accessToken,
+							refresh_token: data.refreshToken,
+						},
 					},
-				},
-				{ upsert: false }
-			);
+					{ upsert: false }
+				);
 		} catch (e) {
 			console.log(e);
 		}
 	}
 
-	async getUsersInfo(userIds: ObjectId[]): Promise<Array<UserWithTokens>> {
+	async getUsersInfo(userIds: ObjectId[]): Promise<Array<UserWithTokens> | undefined> {
 		try {
 			const foundedUsers = await this.usersCollection
 				.aggregate<UserWithTokens>()
